fixed some const bugs in connections fixed mdbsql detection fixed const error in mdb_sql_run_query

+++ src/api/c++/include/sqlrelay/private/bindvar.h
@@ -20,7 +20,7 @@ class bindvar {
 		char	*variable;
 		union {
 			char	*stringval;
-			int32_t	longval;
+			int64_t	longval;
 			struct {
 				double		value;
 				uint32_t	precision;
