+++ src/api/c++/src/cursor/skipandfetch.C
@@ -4,14 +4,14 @@
 #include <config.h>
 #include <sqlrelay/sqlrclient.h>
 
-bool sqlrcursor::skipAndFetch(bool getallrows, uint32_t rowtoget) {
+bool sqlrcursor::skipAndFetch(bool getallrows, uint64_t rowtoget) {
 
 	if (sqlrc->debug) {
 		sqlrc->debugPreStart();
 		sqlrc->debugPrint("Skipping and Fetching\n");
 		if (!getallrows) {
 			sqlrc->debugPrint("	row to get: ");
-			sqlrc->debugPrint((int32_t)rowtoget);
+			sqlrc->debugPrint((int64_t)rowtoget);
 			sqlrc->debugPrint("\n");
 		}
 		sqlrc->debugPreEnd();
@@ -35,7 +35,7 @@ void sqlrcursor::fetchRows() {
 	if (sqlrc->debug) {
 		sqlrc->debugPreStart();
 		sqlrc->debugPrint("Fetching ");
-		sqlrc->debugPrint((int32_t)rsbuffersize);
+		sqlrc->debugPrint((int64_t)rsbuffersize);
 		sqlrc->debugPrint(" rows\n");
 		sqlrc->debugPreEnd();
 	}
@@ -49,7 +49,7 @@ void sqlrcursor::fetchRows() {
 	sqlrc->cs->write(rsbuffersize);
 }
 
-bool sqlrcursor::skipRows(bool getallrows, uint32_t rowtoget) {
+bool sqlrcursor::skipRows(bool getallrows, uint64_t rowtoget) {
 
 	// if we're reading from a cached result set we have to manually skip
 	if (cachesource && cachesourceind) {
@@ -63,9 +63,9 @@ bool sqlrcursor::skipRows(bool getallrows, uint32_t rowtoget) {
 
 		// get the row offset from the index
 		cachesourceind->setPositionRelativeToBeginning(
-				13+sizeof(int32_t)+(rowcount*sizeof(int32_t)));
-		int32_t	rowoffset;
-		if (cachesourceind->read(&rowoffset)!=sizeof(int32_t)) {
+				13+sizeof(int32_t)+(rowcount*sizeof(int64_t)));
+		int64_t	rowoffset;
+		if (cachesourceind->read(&rowoffset)!=sizeof(int64_t)) {
 			setError("The cache file index appears to be corrupt.");
 			return false;
 		}
@@ -77,7 +77,7 @@ bool sqlrcursor::skipRows(bool getallrows, uint32_t rowtoget) {
 
 	// calculate how many rows to skip unless we're buffering the entire
 	// result set or caching the result set
-	uint32_t	skip=0;
+	uint64_t	skip=0;
 	if (rsbuffersize && !cachedest && !getallrows) {
 		skip=(rowtoget-(rowtoget%rsbuffersize))-rowcount; 
 		rowcount=rowcount+skip;
@@ -85,7 +85,7 @@ bool sqlrcursor::skipRows(bool getallrows, uint32_t rowtoget) {
 	if (sqlrc->debug) {
 		sqlrc->debugPreStart();
 		sqlrc->debugPrint("Skipping ");
-		sqlrc->debugPrint((int32_t)skip);
+		sqlrc->debugPrint((int64_t)skip);
 		sqlrc->debugPrint(" rows\n");
 		sqlrc->debugPreEnd();
 	}
