+++ src/connection/sqlrcursor.h
@@ -28,7 +28,7 @@ class bindvar {
 		int16_t	variablesize;
 		union {
 			char	*stringval;
-			int32_t	longval;
+			int64_t	longval;
 			struct	{
 				double		value;
 				uint32_t	precision;
